requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pandas==2.1.2
numpy==1.25.2
scikit-learn==1.3.2
torch==2.1.0
langchain==0.2.17
langchain-community==0.2.19
langchain-aws==0.1.18
langchain-core==0.2.43
streamlit==1.27.2
requests==2.31.0
aws-lambda-powertools==2.36.0
boto3==1.34.131
python-dotenv==1.0.0